% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL.R
\name{as.conn.url}
\alias{as.conn.url}
\alias{as.conn.url.default}
\title{cast to a url() connection}
\usage{
as.conn.url(x, open = "", blocking = TRUE,
  encoding = getOption("encoding"), method, ...)

\method{as.conn.url}{default}(x, open = "", blocking = TRUE,
  encoding = getOption("encoding"), method, ...)
}
\arguments{
\item{x}{an URL object}

\item{open}{see base::url()}

\item{blocking}{see base::url()}

\item{encoding}{see base::url()}

\item{method}{see base::url()}

\item{...}{other arguments passed to specific methods}
}
\value{
a base::connection from url(x$url, ...)
}
\description{
TODO(gp): all 'as.connection' alternatives are needed (as.file, ...)
}

