% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL.R
\name{as.data.frame.URL}
\alias{as.data.frame.URL}
\alias{as.data.frame.URL_list}
\title{coerce an URL to a data.frame}
\usage{
\method{as.data.frame}{URL}(x, row.names = NULL, optional = FALSE,
  RECURSIVE = FALSE, ...)

\method{as.data.frame}{URL_list}(x, row.names = NULL, optional = FALSE,
  RECURSIVE = FALSE, ...)
}
\arguments{
\item{x}{any object}

\item{row.names}{same as as.data.frame}

\item{optional}{same as as.data.frame}

\item{RECURSIVE}{deep expand src URLs}

\item{...}{further arguments}
}
\description{
coerce an URL to a data.frame

see http://www.r-bloggers.com/concatenating-a-list-of-data-frames/
}

