% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL.R
\name{as.URL}
\alias{as.URL}
\alias{as.URL.default}
\title{conver string form of object to URL object}
\usage{
as.URL(x, ...)

\method{as.URL}{default}(x, ...)
}
\arguments{
\item{x}{somthing to be cast to an URL}

\item{...}{other arguments passed to specific methods}
}
\value{
an URL object
}
\description{
conver string form of object to URL object
}

