% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL.R
\name{as.URL_row}
\alias{as.URL_row}
\alias{as.URL_row.default}
\title{conver string form of object to URL object}
\usage{
as.URL_row(x, ...)

\method{as.URL_row}{default}(x, ...)
}
\arguments{
\item{x}{an URL object}

\item{...}{other arguments passed to specific methods}
}
\value{
an named list of data.frame columns
}
\description{
conver string form of object to URL object
}

