% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URLS.R
\docType{methods}
\name{as.URLS}
\alias{as.URLS}
\alias{as.URLS,ANY-method}
\alias{as.URLS,URLS-method}
\alias{as.URLS,list-method}
\title{Change an object to a URLS (collection of ULRs)}
\usage{
as.URLS(x)

as.URLS(x)

\S4method{as.URLS}{URLS}(x)

\S4method{as.URLS}{list}(x)

\S4method{as.URLS}{ANY}(x)
}
\arguments{
\item{x}{Object to be coerced to a duration}
}
\value{
A URLS object
}
\description{
Change an object to a URLS (collection of ULRs)
}
\examples{
\dontrun{
str(us)
}
}
\keyword{chron}
\keyword{classes}
\keyword{manip}
\keyword{methods}

