% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL.R
\name{URL}
\alias{URL}
\title{Create a URL S3 object}
\usage{
URL(url, base = paste0("file://", getwd()), src = NULL, id = NULL,
  meta = c(), ...)
}
\arguments{
\item{url}{string url rapresentation}

\item{base}{a base url for relative urls (string or URL)}

\item{src}{an URL object fron which this object is derived}

\item{id}{entry id (taken from named list/array)}

\item{meta}{an optional array of meta attributes}

\item{...}{other arguments passed to specific methods}
}
\value{
an URL object
}
\description{
Create a URL S3 object
}
\examples{
# simple scalar construction
u <- URL('http://www.w3c.org')
str(u)
as.character(u)

# "derived" URL
u2 <- URL('http://ietf.org', ,src='http://www.w3c.org')
str(u2)
as.character(u2)
as.character(u2$src)

# default schema ('file://') with defaults

s1 <- URL('/etc/R/Renvironment')
s2 <- URL('hosts',base='/etc')
s3 <- URL('.Rhistory')
s4 <- URL('~/.Rprofile')
s5 <- URL('.Rprofile', base='~')

as.character(c(s1,s2,s3,s4,s5))

# construction from a (named) character vector and data.frame cast
l <- c(
    w3c='http://www.w3c.org',
    ietf='http://ietf.org',
    iana='http://www.iana.org',
    hosts='hosts',
    'R env'='/etc/R/Renviron'
 )

 uz <- URL(l, src=URL('https://cran.r-project.org/'), base='/etc')

 df <- as.data.frame.URL(uz)

 head(df,nrow(df))

}

