% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-base.R
\name{URL_init}
\alias{URL_collect}
\alias{URL_config}
\alias{URL_expand}
\alias{URL_fetch}
\alias{URL_init}
\alias{URL_parse}
\title{"Virtual" object initialization}
\usage{
URL_init(x, ...)

URL_config(x, ...)

URL_expand(x, ...)

URL_fetch(x, ...)

URL_parse(x, ...)

URL_collect(x, ...)
}
\arguments{
\item{x}{an URL object}

\item{...}{further arguments}
}
\value{
an list of URL obects

an list of URL obects

an list of URL obects

an list of URL obects

an list of URL obects

an list of URL obects
}
\description{
"Virtual" object initialization

macro substitution

macro substitution

params

params

params
}

