% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL.R
\name{URL_class}
\alias{URL_class}
\title{Assiggn S3 classes to urls}
\usage{
URL_class(url, comp = urltools::url_parse(url), src = NULL, id = NULL,
  classes = NULL, ...)
}
\arguments{
\item{url}{string url rapresentation}

\item{comp}{a parsed url}

\item{src}{source url}

\item{id}{name of url instance (in dotted form)}

\item{classes}{extra classes to include (between schema and id)}

\item{...}{other arguments passed to specific methods}
}
\value{
an array of string to be assigned as S3 classes
}
\description{
Assiggn S3 classes to urls
}

