% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL.R
\name{URL_data}
\alias{URL_data}
\alias{URL_data.default}
\title{extract a persed data object, stored after 'parse'
in object environment}
\usage{
URL_data(x, ...)

\method{URL_data}{default}(x, ...)
}
\arguments{
\item{x}{a URL object}

\item{...}{other arguments passed to specific methods}
}
\value{
an parsed data object
}
\description{
extract a persed data object, stored after 'parse'
in object environment
}

