% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{URL_fetch}
\alias{URL_fetch}
\alias{URL_fetch.default}
\title{Fetch url to local storage}
\usage{
URL_fetch(x, ...)

\method{URL_fetch}{default}(x, ...)
}
\arguments{
\item{x}{an uninitialized URL object}

\item{...}{other arguments passed to specific methods}
}
\value{
a local file:// URL obect or na:// URL in case of failure
}
\description{
Fetch url to local storage
}

