% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL.R
\name{URL_init}
\alias{URL_init}
\alias{URL_init.default}
\title{"Virtual" object initialization}
\usage{
URL_init(x, meta = c(), ...)

\method{URL_init}{default}(x, meta = c(), ...)
}
\arguments{
\item{x}{an uninitialized URL object}

\item{meta}{an optional array of meta attributes}

\item{...}{other arguments passed to specific methods}
}
\value{
an initialized URL obect
}
\description{
"Virtual" object initialization
}

