% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL.R
\name{is.URL_list}
\alias{as.URL_list}
\alias{as.URL_list.list}
\alias{is.URL_list}
\title{check if URL_list object}
\usage{
is.URL_list(x)

as.URL_list(x, ...)

\method{as.URL_list}{list}(x, ...)
}
\arguments{
\item{x}{an arbitrary object or an object of S3 class URL_list}

\item{...}{further arguments}
}
\value{
check condition

an URL_list object
}
\description{
check if URL_list object

throw a conversion error if the list contains other stuff.
}

