% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL.R
\name{URL_meta}
\alias{URL_meta}
\alias{URL_meta.default}
\title{set meta named array argument
as URL object attributes (with prefix 'meta.')}
\usage{
URL_meta(x, meta = c(), ...)

\method{URL_meta}{default}(x, meta = c(), ...)
}
\arguments{
\item{x}{an uninitialized URL object}

\item{meta}{an optional array of meta attributes}

\item{...}{other arguments passed to specific methods}
}
\value{
a 'meta' decorated URL obect
}
\description{
set meta named array argument
as URL object attributes (with prefix 'meta.')
}

