% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL.R
\name{URL_parse}
\alias{URL_parse}
\alias{URL_parse.default}
\title{follows 'R data()' and httr euristics to build a data rapresentation of an URL
the object returned is data.frame or environment}
\usage{
URL_parse(x, ...)

\method{URL_parse}{default}(x, ...)
}
\arguments{
\item{x}{a URL object}

\item{...}{other arguments passed to specific methods}
}
\value{
an list of URL with relerence to internal environment for data object
}
\description{
Additional parameters are passed to read.* function
}

