% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{URL_stop}
\alias{URL_stop}
\alias{URL_stop.default}
\title{error handler for URL functions}
\usage{
URL_stop(x, caller, text, ...)

\method{URL_stop}{default}(x, caller, text, ...)
}
\arguments{
\item{x}{a URL object}

\item{caller}{a caller method name}

\item{text}{error message}

\item{...}{other arguments passed to specific methods}
}
\value{
raises a stop condition
}
\description{
error handler for URL functions
}

