% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URLS.R
\docType{class}
\name{URLS-class}
\alias{URLS}
\alias{URLS-class}
\alias{format.URLS}
\alias{is.URLS}
\alias{show,URLS-method}
\title{URLS class}
\usage{
\S4method{show}{URLS}(object)

\method{format}{URLS}(x, ...)

is.URLS(x)
}
\arguments{
\item{object}{an Object}

\item{x}{an Object}

\item{...}{further arguments}
}
\description{
URL Collection (list)
Store a List of URLs as a list slot
}
\details{
URLS are used as promary parameter and return value from URL transormation methods.
}

