% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls-methods.R
\docType{methods}
\name{us}
\alias{us}
\alias{us,ANY-method}
\alias{us,URLS-method}
\title{create a list of URL fron scala or list(vector arguments}
\usage{
us(x, ...)

us(x, ...)

\S4method{us}{URLS}(x, ...)

\S4method{us}{ANY}(x, ...)
}
\arguments{
\item{x}{a scalar/list parameter to URL constructor}

\item{...}{further arguments}
}
\value{
A URLS object
}
\description{
Invoke URL constructor, initialize urls and the wrao in URLS object
}

