% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls-methods.R
\docType{methods}
\name{ux}
\alias{ux}
\alias{ux,ANY-method}
\alias{ux,URLS-method}
\title{create a list of URL fron scala or list(vector arguments}
\usage{
ux(x, f, ...)

ux(x, f, ...)

\S4method{ux}{URLS}(x, f, ...)

\S4method{ux}{ANY}(x, f, ...)
}
\arguments{
\item{x}{a scalar/list parameter to URL constructor}

\item{f}{an URL tranformation function to apply}

\item{...}{further arguments}
}
\value{
A URLS object
}
\description{
Invoke URL constructor, initialize urls and the wrao in URLS object
}

