% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL.R
\name{URL}
\alias{URL}
\alias{as.URL}
\alias{as.URL.default}
\alias{as.URL_row}
\alias{as.URL_row.default}
\alias{as.character.URL}
\alias{as.data.frame.URL}
\alias{is.URL}
\title{Create a URL S3 object}
\usage{
URL(url, base = paste0("file://", getwd()), src = NULL, id = NULL,
  meta = c(), ...)

is.URL(x)

as.URL(x, ...)

\method{as.URL}{default}(x, ...)

\method{as.character}{URL}(x, ...)

as.URL_row(x, ...)

\method{as.URL_row}{default}(x, ...)

\method{as.data.frame}{URL}(x, row.names = NULL, optional = FALSE,
  RECURSIVE = FALSE, ...)
}
\arguments{
\item{url}{string url rapresentation}

\item{base}{a base url for relative urls (string or URL)}

\item{src}{an URL object fron which this object is derived}

\item{id}{entry id (taken from named list/array)}

\item{meta}{an optional array of meta attributes}

\item{...}{other arguments passed to specific methods}

\item{x}{an arbitrary object or an object of S3 class URL}

\item{row.names}{same as as.data.frame}

\item{optional}{same as as.data.frame}

\item{RECURSIVE}{deep expand src URLs}
}
\value{
an URL object

check condition

an URL object

an named list of data.frame columns
}
\description{
Create a URL S3 object

check if URL object

conver as.character() string from of object to URL object

conver string form of object to URL object

coerce an URL vector to a data.frame
}
\examples{
# simple scalar construction
u <- URL('http://www.w3c.org')
str(u)
as.character(u)

# "derived" URL
u2 <- URL('http://ietf.org', ,src='http://www.w3c.org')
as.character(u2)
as.character(u2$src)

# default schema ('file://') with defaults

s1 <- URL('/etc/R/Renvironment')
s2 <- URL('hosts',base='/etc')
s3 <- URL('.Rhistory')
s4 <- URL('~/.Rprofile')
s5 <- URL('.Rprofile', base='~')

sapply(list(s1,s2,s3,s4,s5), as.character)

# construction from a (named) character vector and data.frame cast
l <- c(
    w3c='http://www.w3c.org',
    ietf='http://ietf.org',
    iana='http://www.iana.org',
    hosts='hosts',
    'R env'='/etc/R/Renviron'
 )

 uz <- URL(l, src=URL('https://cran.r-project.org/'), base='/etc')

 df <- as.data.frame.URL(uz)

 head(df,nrow(df))

}

