% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{URL_stop_unsupported}
\alias{URL_stop_unsupported}
\alias{URL_stop_unsupported.default}
\title{unsupported operation error}
\usage{
URL_stop_unsupported(x, caller, ...)

\method{URL_stop_unsupported}{default}(x, caller, ...)
}
\arguments{
\item{x}{a URL object}

\item{caller}{a caller method name}

\item{...}{other arguments passed to specific methods}
}
\value{
raises a stop condition
}
\description{
unsupported operation error
}

